storage: split memory and autospill upsert impls into their own files
